rename, adapt and move stxxl files to core
