--- extlib/stxxl/include/stxxl/bits/namespace.h
@@ -1,20 +1,0 @@
-/***************************************************************************
- *  include/stxxl/bits/namespace.h
- *
- *  Part of the STXXL. See http://stxxl.sourceforge.net
- *
- *  Copyright (C) 2007 Roman Dementiev <dementiev@ira.uka.de>
- *
- *  Distributed under the Boost Software License, Version 1.0.
- *  (See accompanying file LICENSE_1_0.txt or copy at
- *  http://www.boost.org/LICENSE_1_0.txt)
- **************************************************************************/
-
-#ifndef STXXL_NAMESPACE_HEADER
-#define STXXL_NAMESPACE_HEADER
-
-#define STXXL_BEGIN_NAMESPACE namespace stxxl {
-#define STXXL_END_NAMESPACE \
-    }
-
-#endif // !STXXL_NAMESPACE_HEADER
